Extract profile button to main addon
